Got package to work, made proper singleton, implemented standard interface for packages that can render
