+++ packages/jetpack-common-interfaces/src/Renderable_Interface.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Renderable Interface
+ *
+ * @package jetpack-common-interfaces
+ */
+
+namespace Jetpack\Assets;
+
+interface Renderable_Interface {
+
+	/**
+	 * Returns rendered package output
+	 *
+	 * @return string
+	 */
+	public static function render();
+}
