+++ packages/jetpack-logo/composer.json
@@ -3,7 +3,7 @@
 	"description": "A logo for Jetpack",
 	"type": "library",
 	"license": "GPL-2.0-or-later",
-	"version": "1.0.0",
+	"version": "1.0.1",
 	"require": {},
 	"autoload": {
 		"psr-4": {
