+++ packages/jetpack-logo/src/README.md
